+++ packages/builder/webpack-builder/src/types/plugin.ts
@@ -4,11 +4,13 @@ import type {
   OnExitFn,
   OnAfterBuildFn,
   OnBeforeBuildFn,
+  OnDevCompileDoneFn,
   ModifyWebpackChainFn,
   ModifyWebpackConfigFn,
   ModifyBuilderConfigFn,
   OnAfterCreateCompilerFn,
   OnBeforeCreateCompilerFn,
+  OnAfterStartDevServerFn,
   OnBeforeStartDevServerFn,
 } from './hooks';
 
@@ -28,8 +30,10 @@ export type BuilderPluginAPI = {
   onExit: (fn: OnExitFn) => void;
   onAfterBuild: (fn: OnAfterBuildFn) => void;
   onBeforeBuild: (fn: OnBeforeBuildFn) => void;
+  onDevCompileDone: (fn: OnDevCompileDoneFn) => void;
   onAfterCreateCompiler: (fn: OnAfterCreateCompilerFn) => void;
   onBeforeCreateCompiler: (fn: OnBeforeCreateCompilerFn) => void;
+  onAfterStartDevServer: (fn: OnAfterStartDevServerFn) => void;
   onBeforeStartDevServer: (fn: OnBeforeStartDevServerFn) => void;
 
   // Modifiers