+++ packages/server/prod-server/tests/context.test.ts
@@ -3,6 +3,8 @@ import { Readable } from 'stream';
 import httpMocks from 'node-mocks-http';
 import { createContext } from '../src/libs/context';
 
+const Etag = 'W/"c8e8-KIdhFJWJkiBDAQ+6qEnlCiVlE7c"';
+
 describe('test server context', () => {
   test('should route api work correctly', () => {
     const req = httpMocks.createRequest({
@@ -48,5 +50,50 @@ describe('test server context', () => {
         name: 'foo',
       },
     });
+  });
+
+  test('should fresh work correctly', () => {
+    const req = httpMocks.createRequest({
+      url: '/',
+      headers: {},
+      eventEmitter: Readable,
+      method: 'GET',
+    });
+    const res = httpMocks.createResponse({ eventEmitter: EventEmitter });
+    const context = createContext(req, res);
+    expect(context.fresh).toBeFalsy();
+
+    req.headers['if-none-match'] = Etag;
+    expect(context.fresh).toBeFalsy();
+
+    res.setHeader('etag', Etag);
+    expect(context.fresh).toBeTruthy();
+
+    res.setHeader('etag', Etag.replace('c', 'd'));
+    expect(context.fresh).toBeFalsy();
+
+    res.setHeader('etag', Etag);
+    delete req.headers['if-none-match'];
+    expect(context.fresh).toBeFalsy();
+  });
+
+  test('should etag work correctly', () => {
+    const req = httpMocks.createRequest({
+      url: '/',
+      headers: {},
+      eventEmitter: Readable,
+      method: 'GET',
+    });
+    const res = httpMocks.createResponse({ eventEmitter: EventEmitter });
+    const context = createContext(req, res, { etag: true });
+
+    res.send('hello jupiter');
+    expect(context.status).toBe(200);
+    const etag = context.res.getHeader('ETag');
+    expect(etag).toBeDefined();
+
+    req.headers['if-none-match'] = etag as string;
+    res.send('hello jupiter');
+    expect(context.status).toBe(304);
   });
 });