+++ scripts/prebundle/src/constant.ts
@@ -191,8 +191,15 @@ export const TASKS: TaskConfig[] = [
     packageDir: 'builder/webpack-builder',
     packageName: '@modern-js/webpack-builder',
     dependencies: [
+      'open',
       'tapable',
       'webpack-merge',
+      {
+        name: 'webpack-5-chain',
+        externals: {
+          tapable: '../tapable',
+        },
+      },
       {
         name: 'webpack-sources',
         ignoreDts: true,