+++ tests/integration/webpack-builder/scripts/shared.ts
@@ -1,5 +1,4 @@
 import { join } from 'path';
-import { InspectorWebpackPlugin } from '@modern-js/inspector-webpack-plugin';
 
 export const createBuilder = async () => {
   const { createBuilder } = await import('@modern-js/webpack-builder');
@@ -9,20 +8,12 @@ export const createBuilder = async () => {
       main: join(process.cwd(), 'src', 'index.ts'),
     },
     configPath: __filename,
-    builderConfig: {},
-  });
-
-  builder.addPlugins([
-    {
-      name: 'inspect-plugin',
-      setup(api) {
-        api.modifyWebpackConfig(config => {
-          // Webpack devtool
-          config.plugins?.push(new InspectorWebpackPlugin());
-        });
+    builderConfig: {
+      tools: {
+        inspector: {},
       },
     },
-  ]);
+  });
 
   return builder;
 };